Initial version
Initial version of basic utility script.
Note that this script was created for specific internal needs as laid out in the
use cases of the readme. It is also in an alpha state. It may not do what you
want, and it also may not work for your need. Feel free to submit pull requests
if you see any problems or have any enhancements.